vim: Fix bug where fzf doesn't update working directory
When you navigate to another file with fzf, it doesn't update the
current working directory.

This problem was caused by how we configured fzf to perform its searches
starting from the project root, namely by using the ProjectRootExe
command from the vim-projectroot plugin. However, what ProjectRootExe
actually does is change the working directory to the project root before
executing the command it is given.

The solution is to, instead, pass the project root path to fzf's
commands as an argument. The Files command already accepts a path as an
argument, but not Ag, which requires us to override the definition of
the command itself.

This patch also contains a couple additional fixes:
* The keymap for 'Ag!' will type 'Ag!' in the command prompt but not
  execute it, allowing for us to execute with a search pattern by typing
  the pattern and manually executing the command instead
* Bugfix for comments that are in-line with keymapping declarations,
  see https://stackoverflow.com/a/24717020