shell: Convert shell functions to scripts
Previously, the functions were executed using whatever shell is
currently in use. For example, if the function 'v' is invoked while in
zsh, then the function will be interpreted by the zsh shell.

However, as it turns out, a built-in command (i.e. 'read') used by one
function behaves differently when interpreted by bash vs when
interpreted by zsh. This led to inconsistent behaviour when invoking the
function based on which shell is currently in use.

By converting the functions to scripts with the bash shebang on top, we
effectively ensure that the logic is always interpreted by bash, leading
to consistent behaviour whether the script is invoked in bash or in zsh.

We do this conversion for all functions for consistency reasons and to
ensure that no similar problem arises in the future.