git: Make git_reset_hard.sh compatible with older versions of git
In older versions of git, 'git add -A' stages only files within the
current directory. This means that if you were to invoke this command
from within a subdirectory in a git project, files that are not in that
subdirectory but are otherwise in that git project (e.g. files in the
project root) are not staged. To stage all files within the entire git
project, you need to specify ':/' (i.e. 'git add -A :/').

Newer versions of git do not require the ':/' and stage all git project
files by default.

Source: https://stackoverflow.com/a/33449169