git: Make git_push_force.sh safer
This patch removes the ability to specify a remote branch to push a
local branch to, and instead allows for pushing only to a remote branch
with the same name.

This patch also generally makes git_push_force.sh safer by adding checks
to prevent fat-finger errors and allowing for only expected use-cases.