shell, tmux: Fix bug in v and vs where pane directory wasn't being updated
Previously: suppose you launch vim through one of the helper scripts, v
or vs. If you were to navigate to a file in a different directory and
then open a new tmux pane, the pane's directory is the same directory
wherein you originally executed the helper script rather than the
directory of the current vim buffer.

This is because the helper script, which is executed as its own child
process, spawns a whole new child process of its own when it invokes the
'vim' command. Importantly, from the point-of-view of tmux, the process
running in the current pane is the helper script, but it does not see
the child process actually running vim itself. Additionally, even if the
vim process changes its current directory, the current directory of the
script process is unaffected. Therefore, when you open a new tmux pane
with the '-c {#pane_current_path}' argument, it uses the current
directory of the script process rather than that of vim.

This patch fixes the above issue by changing the helper scripts to
execute vim with the 'exec' command. This command replaces the current
process with the given command instead of spawning a child process.

Source: https://www.computerhope.com/unix/bash/exec.htm