vim: Fix bug: FindRootDirectory() returns empty string when in buffer opened by :FileBeagle
This patch applies a workaround to fix an issue where the
FindRootDirectory() helper function provided by vim-rooter returns an
empty string when invoked in a buffer opened by the :FileBeagle command.

This issue led to keymaps which invoke FZF to not work correctly (i.e.
not perform searches from the project root when configured to do so).

However, for some reason, FindRootDirectory() works just fine if the
FileBeagle buffer was opened by invoking :edit on the current working
directory instead of using the :FileBeagle command. This finding is the
basis for the workaround made in this patch.