zsh: Fix _arguments:comparguments:325 errors on zsh start-up
This commit fixes an issue where the following error message would show
up whenever zsh starts up [1]:

    _arguments:comparguments:325: can only be called from completion
    function

This commit fixes the issue by replacing the zgen repo used by our
dotfiles with a fork that contains a community-contributed fix [2].

This is done since the official zgen repo appears to no longer be
actively maintained [3], meaning that the community-contributed PR is
unlikely to be accepted any time soon.

In the long-term, we should consider migrating to actively maintained
forks of zgen like zgenom [4] or a different zsh plugin manager.
However, given that zsh plugin managers have a tendency to go inactive,
it might be best to consider finding a way to remove our reliance on
plugin managers entirely and just clone the zsh plugins directly.

[1] https://github.com/zsh-users/zsh-completions/issues/809
[2] https://github.com/tarjoilija/zgen/pull/132
[3] https://github.com/tarjoilija/zgen/issues/123
[4] https://github.com/tarjoilija/zgen/issues/123#issuecomment-733178282