vim: Fix issue where 'jump to definition' wasn't working in Go integration test files
There was an issue where hitting "Ctrl + ]" on symbols (e.g. functions)
in integration test files would not work and would instead yield the
error:

    vim-go: no packages returned: packages.Load error

This was because gopls was unable to load packages used by integration
test files, which could be verified using:

    $ gopls -rpc.trace -v check <file.go>
    ...
    gopls: no packages returned: packages.Load error

This was because gopls needed to be told to include integration test
files when parsing Go files using '-tags integration' [1], similar to
how you need to set the same flag for 'go build' / 'go test' to
build/test integration test files.

To fix the issue, we could have added '-tags integration' to GOFLAGS,
but to isolate the change to just vim, we set vim-go's g:go_build_tags
instead.

For more info on Go build tags, what they're for, and how they work, see
https://wawand.co/blog/posts/using-build-tags/.

[1] https://github.com/golang/go/issues/41081#issuecomment-682269983